requirements.txt
Django==4.2.16
djangorestframework==3.15.2
mysqlclient==2.2.7
gunicorn==21.2.0
uvicorn==0.34.0
python-dotenv==1.0.1
requests==2.32.3
django_cron==0.6.0
pytz==2024.2
django-redis==5.4.0
PyJWT==2.10.1
drf-spectacular==0.28.0
whitenoise==6.7.0
django-widget-tweaks==1.5.0
isodate==0.7.2
boto3==1.35.97
Pillow==11.1.0